Supports pre-query functionality If the same question is asked to a certain robot, the answer is automatically read from localStorage instead of making a new request Saves time
