+++ src/utils/index.js
@@ -1,0 +1,1 @@
+export { get_messages } from './storage'
